SDK-1119: Implement Source based constraints for dynamic scenarios
To allow for other types of constraint, and optional parameters in the
future the helper methods for dynamic policy construction now accept
variadic interface{} parameters. A type switch statement is used to
process the slice based on the type of the parameters. At present,
SourceConstraint pointers will be added to the wanted attribute.

This behavior has been implemented for bother normal and derived
attributes.